Solve C/D in Biweekly contest 126
C. Took a minute to figure out. Overall idea is that the exact character
placement is not important, but rather it's the character count that has
an impact on the answer. Solution tries to evenly insert all characters
to get their count as close to equal as possible. Code isn't very good
- it's an opportunity to learn more idiomatic Kotlin.

D. Straightforward DP. O(n^3). Implementation is tricky because we need
to ensure that state updates don't overlap. Spent far too long
identifying modulus issues.